% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\docType{data}
\name{daily}
\alias{daily}
\title{Classic Canadian weather data set.}
\format{
An object containing temperature and precipitation data for 35 Canadian cities.
}
\usage{
data(daily)
}
\description{
Classic Canadian weather data set.
}
\references{
Ramsay J, Hooker G, Graves S (2009). \emph{Functional Data Analysis with R and MATLAB}. Springer-Verlag, New York. ISBN 9780387981857.
}
\keyword{datasets}
