% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecator.R
\docType{data}
\name{tecator}
\alias{tecator}
\title{Classic Tecator data set.}
\format{
An object containing the response and absorbance curve values.
}
\usage{
data(tecator)
}
\description{
Classic Tecator data set.
}
\references{
Thodberg, H. H. (2015). “Tecator meat sample dataset.” StatLib Datasets Archive. http://lib.stat.cmu.edu/datasets/tecator.
}
\keyword{datasets}
