% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\docType{data}
\name{daily}
\alias{daily}
\title{Classic Canadian weather data set.}
\format{
An object containing temperature and precipitation data for 35 Canadian cities.
}
\usage{
data(daily)
}
\description{
Classic Canadian weather data set.
}
\references{
Reference can be found in "Deep Learning with Functional Inputs".
}
\keyword{datasets}
