% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFuncNN.R
\docType{package}
\name{FuncNN-package}
\alias{FuncNN}
\alias{FuncNN-package}
\title{FuncNN: Functional Neural Networks}
\description{
A collection of functions which fit functional neural network models. In
            other words, this package will allow users to build deep learning models 
            that have either functional or scalar responses paired with functional and 
            scalar covariates. We implement the theoretical discussion found 
            in (Thind, Multani and Cao, 2020) through the help of a main fitting and 
            prediction function as well as a number of helper functions to help with 
            cross-validation, tuning, and displaying estimated functional weights.
}
\references{
Thind B, Multani K, Cao J (2020). “Deep Learning with Functional Inputs.” 2006.09590. https://arxiv.org/abs/2006.09590.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arxiv.org/abs/2006.09590}
  \item \url{https://github.com/b-thi/FuncNN}
}

}
\author{
\strong{Maintainer}: Barinder Thind \email{barinder.thi@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Jiguo Cao
}

Other contributors:
\itemize{
  \item Richard Groenewald [contributor]
  \item Sidi Wu [contributor]
}

}
\keyword{internal}
