% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecator.R
\docType{data}
\name{tecator}
\alias{tecator}
\title{Classic Tecator data set.}
\format{
An object containing the response and absorbance curve values.
}
\usage{
data(tecator)
}
\description{
Classic Tecator data set.
}
\references{
Reference can be found in "Deep Learning with Functional Inputs".
}
\keyword{datasets}
